# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "WebApiEngine",
    "version": "1.0"
  },
  "paths": {
...
    "/Form": {
      "post": {
        "tags": [
          "Form"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/FormDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              }
            }
          }
        }
      }
    },
    "/Form/compatibility": {
      "get": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "needsMigration",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormCompatibilityItemDtoArrayApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormCompatibilityItemDtoArrayApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormCompatibilityItemDtoArrayApiStatusResult"
                }
              }
            }
          }
        }
      }
    },
    "/Form/{formId}/draft": {
      "get": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormAuthoringDraftRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormAuthoringDraftRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormAuthoringDraftRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expectedRevision",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Form/{formId}/publish": {
      "post": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormAuthoringDraftRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormAuthoringDraftRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormAuthoringDraftRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Form/{formId}/preview": {
      "post": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewFormAuthoringRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewFormAuthoringRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/PreviewFormAuthoringRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringPreviewDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringPreviewDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringPreviewDtoApiStatusResult"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/Form/{formId}/{formIdentifier}": {
      "get": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "formIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              }
            }
          }
        }
      }
    },
    "/Form/meta/{formId}": {
      "get": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormMetaDataDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormMetaDataDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/FormMetaDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoApiStatusResult"
                }
              }
            }
          }
        }
      }
    },
    "/Form/meta": {
      "get": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoArrayApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoArrayApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetaDataDtoArrayApiStatusResult"
                }
              }
            }
          }
        }
      }
    },
    "/Form/result": {
      "post": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTaskResultDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTaskResultDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UserTaskResultDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FormAuthoringDraftDtoApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "$ref": "#/components/schemas/FormAuthoringDraftDto"
          }
        },
        "additionalProperties": false
      },
...
      "FormCompatibilityItemDtoArrayApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FormCompatibilityItemDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FormDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "formId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "version": {
            "$ref": "#/components/schemas/VersionDto"
          },
          "formData": {
            "type": "string",
            "nullable": true
          },
          "validationProfile": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FormDtoApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "$ref": "#/components/schemas/FormDto"
          }
        },
        "additionalProperties": false
      },
...
      "PreviewFormAuthoringRequestDto": {
        "required": [
          "formData"
        ],
        "type": "object",
        "properties": {
          "formData": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PublishFormAuthoringDraftRequestDto": {
        "required": [
          "expectedRevision"
        ],
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "SaveFormAuthoringDraftRequestDto": {
        "required": [
          "expectedRevision",
          "formData"
        ],
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          },
          "formData": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}