# src/WebApiEngine/Controller/FormController.cs
using WebApiEngine.BusinessLogic;
using WebApiEngine.Mappers;
using WebApiEngine.Shared;
using WebApiEngine.Auth;
using Microsoft.AspNetCore.Authorization;

namespace WebApiEngine.Controller;

[ApiController, Route("[controller]")]
public class FormController(
    IStorageSystem storageSystem,
    FormBusinessLogic formBusinessLogic,
    BpmnBusinessLogic bpmnBusinessLogic,
    UserTaskCompletionService completionService,
    FormAuthoringService authoringService,
    FormCompatibilityService compatibilityService): ControllerBase
{

    [HttpPost()]
    [Authorize(Policy = FlowzerPolicies.Modeler)]
    public async Task<ActionResult<ApiStatusResult<FormDto>>> SaveForm(FormDto formDto)
    {
        Form form;
        try
        {
            form = formDto.ToModel();
        }
        catch (ArgumentException e)
        {
            return BadRequest(new ApiStatusResult<FormDto>(e.Message));
        }
        
        if (form.FormId == Guid.Empty)
            return BadRequest(new ApiStatusResult<FormDto>("FormId is required"));
        
        form = await formBusinessLogic.SaveForm(form);
        
        var retForm = form.ToDto();
        
        return Ok(new ApiStatusResult<FormDto>()
        {
            Result = retForm,
            Successful = true,
        });
    }

    /// <summary>Prueft veroeffentlichte Formularversionen und den Autorenentwurf.</summary>
    [HttpGet("compatibility")]
    [Authorize(Policy = FlowzerPolicies.Modeler)]
    [ProducesResponseType<ApiStatusResult<FormCompatibilityItemDto[]>>(StatusCodes.Status200OK)]
    public async Task<ActionResult<ApiStatusResult<FormCompatibilityItemDto[]>>> GetCompatibility(
        [FromQuery] bool? needsMigration = null)
    {
        var result = await compatibilityService.GetAsync(needsMigration);
        return Ok(new ApiStatusResult<FormCompatibilityItemDto[]>(result.ToArray()));
    }

    /// <summary>Liest den gemeinsamen Entwurf oder eine unveraenderliche Veroeffentlichungsbasis.</summary>
    [HttpGet("{formId:guid}/draft")]
    [Authorize(Policy = FlowzerPolicies.Modeler)]
    [ProducesResponseType<ApiStatusResult<FormAuthoringDraftDto>>(StatusCodes.Status200OK)]
    [ProducesResponseType<ApiStatusResult<FormAuthoringDraftDto>>(StatusCodes.Status404NotFound)]
    public async Task<ActionResult<ApiStatusResult<FormAuthoringDraftDto>>> GetDraft(Guid formId) =>
        Ok(new ApiStatusResult<FormAuthoringDraftDto>(await authoringService.GetAsync(formId)));

    /// <summary>Speichert den Autorenentwurf nur bei passender Revision.</summary>
    [HttpPut("{formId:guid}/draft")]
    [Authorize(Policy = FlowzerPolicies.Modeler)]
    [ProducesResponseType<ApiStatusResult<FormAuthoringDraftDto>>(StatusCodes.Status200OK)]
    [ProducesResponseType<WebApiEngine.Middleware.ApiProblemDetails>(StatusCodes.Status409Conflict, "application/problem+json")]
    public async Task<ActionResult<ApiStatusResult<FormAuthoringDraftDto>>> SaveDraft(
        Guid formId,
        SaveFormAuthoringDraftRequestDto request) =>
        Ok(new ApiStatusResult<FormAuthoringDraftDto>(await authoringService.SaveAsync(formId, request)));

    /// <summary>Verwirft den Autorenentwurf nur bei passender Revision.</summary>
    [HttpDelete("{formId:guid}/draft")]
    [Authorize(Policy = FlowzerPolicies.Modeler)]
    [ProducesResponseType(StatusCodes.Status204NoContent)]
    [ProducesResponseType<WebApiEngine.Middleware.ApiProblemDetails>(StatusCodes.Status409Conflict, "application/problem+json")]
    public async Task<IActionResult> DeleteDraft(Guid formId, [FromQuery] long expectedRevision)
    {
        await authoringService.DeleteAsync(formId, expectedRevision);
        return NoContent();
    }

    /// <summary>Prueft und veroeffentlicht genau den erwarteten Entwurfsstand.</summary>
    [HttpPost("{formId:guid}/publish")]
    [Authorize(Policy = FlowzerPolicies.Modeler)]
    [ProducesResponseType<ApiStatusResult<FormDto>>(StatusCodes.Status200OK)]
    [ProducesResponseType<WebApiEngine.Middleware.ApiValidationProblem>(StatusCodes.Status422UnprocessableEntity, "application/problem+json")]
    [ProducesResponseType<WebApiEngine.Middleware.ApiProblemDetails>(StatusCodes.Status409Conflict, "application/problem+json")]
    public async Task<ActionResult<ApiStatusResult<FormDto>>> PublishDraft(
        Guid formId,
        PublishFormAuthoringDraftRequestDto request) =>
        Ok(new ApiStatusResult<FormDto>((await authoringService.PublishAsync(formId, request.ExpectedRevision)).ToDto()));

    /// <summary>Loest einen lokalen Autorenstand serverseitig auf, ohne ihn zu speichern.</summary>
    [HttpPost("{formId:guid}/preview")]
    [Authorize(Policy = FlowzerPolicies.Modeler)]
    [ProducesResponseType<ApiStatusResult<FormAuthoringPreviewDto>>(StatusCodes.Status200OK)]
    [ProducesResponseType<WebApiEngine.Middleware.ApiValidationProblem>(StatusCodes.Status422UnprocessableEntity, "application/problem+json")]
    public async Task<ActionResult<ApiStatusResult<FormAuthoringPreviewDto>>> PreviewDraft(
        Guid formId,
        PreviewFormAuthoringRequestDto request) =>
        Ok(new ApiStatusResult<FormAuthoringPreviewDto>(
            await authoringService.PreviewAsync(formId, request.FormData)));

    [HttpGet("{formId}/{formIdentifier}")]
    public async Task<ActionResult<ApiStatusResult<FormDto>>> GetForm(Guid formId, string formIdentifier)
    {
        var allVersions = (await storageSystem.FormStorage.GetForms(formId)).ToList();
        if (formIdentifier == "latest")
        {
            if (allVersions.Count == 0)
                return NotFound(new ApiStatusResult<FormDto>(){Successful = false, ErrorMessage = "formular has no versions"});
            formId = allVersions.OrderByDescending(x => x.Version).First().Id;
        }
        else
        {
            Model.Version versionFromFormIdentifier;
            try
            {
                versionFromFormIdentifier = Model.Version.FromString(formIdentifier);
            }
            catch (ArgumentException e)
            {
                return BadRequest(new ApiStatusResult<FormDto>(e.Message));
            }

            // Der API-Pfad arbeitet mit der stabilen FormId, gespeichert werden die Versionen aber
            // unter ihrer konkreten Formular-Instanz-ID. Deshalb wird hier zuerst die Zielversion aufgelöst.
            var foundVersion =  allVersions.SingleOrDefault(x => x.Version.Equals(versionFromFormIdentifier));
            if (foundVersion == null)
                return NotFound(new ApiStatusResult<FormDto>(){Successful = false, ErrorMessage = "Version not found"});

            formId = foundVersion.Id;
        }

        var formMetaData = await storageSystem.FormStorage.GetForm(formId);
        return Ok(new ApiStatusResult<FormDto>(formMetaData.ToDto()));
    }
    
    
    #region MetaData

        [HttpGet("meta/{formId}")]
        public async Task<ActionResult<ApiStatusResult<FormMetaDataDto>>> GetFormMetadata(Guid formId)
        {
            var formMetaData = await storageSystem.FormStorage.GetFormMetaData(formId);
            return Ok(new ApiStatusResult<FormMetaDataDto>(formMetaData.ToDto()));
        }
        
        [HttpGet("meta")]
        public async Task<ActionResult<ApiStatusResult<FormMetaDataDto[]>>> GetFormMetadatas([FromQuery] string? search)
        {
            
            var formMetaData = await storageSystem.FormStorage.GetFormMetadatas();
            if (!string.IsNullOrEmpty(search))
                formMetaData = formMetaData.Where(x => string.Compare(x.Name,search, StringComparison.InvariantCultureIgnoreCase) == 0).ToList();
            return Ok(new ApiStatusResult<FormMetaDataDto[]>(formMetaData.Select(metadata => metadata.ToDto()).ToArray()));
        }
        
        [HttpPost("meta/{formId}")]
        [Authorize(Policy = FlowzerPolicies.Modeler)]
        public async Task<ActionResult<ApiStatusResult>> SaveFormMetadata(Guid formId, FormMetaDataDto formMetadataDto)
        {
            formMetadataDto.FormId = formId;
            
            await storageSystem.FormStorage.SaveFormMetaData(formMetadataDto.ToModel());
            return Ok(new ApiStatusResult(){Successful = true});
        }

        /// <summary>
        /// Entfernt ein Formular samt allen seinen Versionen.
        ///
        /// Benutzt ein deployter Workflow das Formular, bleibt es stehen: Ein Aufgabenformular
        /// wird ueber seinen <em>Namen</em> aufgeloest, ein geloeschtes liesse jede Aufgabe
        /// dieses Schrittes mit „No form named …" stehen — auch in bereits laufenden Instanzen.
        /// Der Aufrufer bekommt die betroffenen Workflows genannt, statt vor einem
        /// unerklaerlichen Fehlschlag zu stehen.
        /// </summary>
        [HttpDelete("meta/{formId:guid}")]
        [Authorize(Policy = FlowzerPolicies.Modeler)]
        [ProducesResponseType<ApiStatusResult<FormMetaDataDto>>(StatusCodes.Status200OK)]
        [ProducesResponseType<ApiStatusResult<FormMetaDataDto>>(StatusCodes.Status404NotFound)]
        [ProducesResponseType<ApiStatusResult<FormMetaDataDto>>(StatusCodes.Status409Conflict)]
        public async Task<ActionResult<ApiStatusResult<FormMetaDataDto>>> DeleteFormMetadata(Guid formId)
        {
            // Suchen, pruefen und loeschen liegen in der Engine unter derselben Sperre wie das
            // Deployen; hier auseinandergezogen koennte dazwischen deployt oder umbenannt werden.
            var ergebnis = await bpmnBusinessLogic.DeleteFormIfUnused(formId);

            if (ergebnis.Metadata is null)
            {
                return NotFound(new ApiStatusResult<FormMetaDataDto>(
                    errorMessage: $"Es gibt kein Formular mit der Kennung {formId}."));
            }

            if (ergebnis.UsedBy.Count > 0)
            {
                return Conflict(new ApiStatusResult<FormMetaDataDto>(
                    errorMessage: $"Das Formular \u201e{ergebnis.Metadata.Name}\u201c wird von "
                                  + $"{string.Join(", ", ergebnis.UsedBy)} benutzt. Erst dort ersetzen, dann loeschen."));
            }

            // Bewusst der geloeschte Eintrag: `new ApiStatusResult<string>(...)` traefe den
            // Konstruktor fuer die Fehlermeldung und meldete den Erfolg als Fehlschlag.
            return Ok(new ApiStatusResult<FormMetaDataDto>(ergebnis.Metadata.ToDto()));
        }

    #endregion


    #region MessageHandling

        
    [HttpPost("result")]
    [ProducesResponseType<ApiStatusResult>(StatusCodes.Status200OK)]
    [ProducesResponseType<ApiStatusResult>(StatusCodes.Status400BadRequest)]
    [ProducesResponseType(typeof(void), StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<ApiStatusResult>(StatusCodes.Status404NotFound)]
    [ProducesResponseType<WebApiEngine.Middleware.ApiValidationProblem>(StatusCodes.Status422UnprocessableEntity, "application/problem+json")]
    [ProducesResponseType<WebApiEngine.Middleware.ApiProblemDetails>(StatusCodes.Status409Conflict, "application/problem+json")]
    public async Task<ActionResult<ApiStatusResult>> HandleUserFormData(
        UserTaskResultDto formMetadataDto,
        [FromHeader(Name = WebApiEngine.Idempotency.HttpIdempotency.HeaderName)] string? _idempotencyKey = null)
    {
        try
        {
            var data = formMetadataDto.ToModel();
            var outcome = await completionService.CompleteAsync(data);
            return outcome == UserTaskCompletionOutcome.Completed
                ? Ok(new ApiStatusResult { Successful = true })
                : NotFound(new ApiStatusResult("The user task was not found."));
        }
        catch (ArgumentException e)
        {
            return BadRequest(new ApiStatusResult(e.Message));
        }
    }

    #endregion
    

}